(#7) Separate GraphicsEnvironment content
New Additions:
All static methods dealing with `GraphicsEnvironment` have been moved `Display → DisplayUtil`.
